+++ web/js/apps/repository/importers/PDFMetadataTest.ts
@@ -1,8 +1,10 @@
 import {PDFMetadata} from './PDFMetadata';
 import {Files} from '../../../../../web/js/util/Files';
 import {HitMap} from '../../../util/HitMap';
+import {Strings} from '../../../util/Strings';
 
 describe('PDF Metadata', function() {
+    this.timeout(999999);
 
     xit("basic", async function() {
         const pdfMeta = await PDFMetadata.getMetadata("/home/burton/Downloads/SSRN-id2594754.pdf");
@@ -11,11 +13,13 @@ describe('PDF Metadata', function() {
     });
 
 
-    xit("build property index", async function() {
+    it("build property index", async function() {
 
         const hitMap = new HitMap();
 
-        await Files.recursively("/home/burton/.polar/stash", async (path) => {
+        let nrFiles: number = 0;
+
+        await Files.recursively("/d0/polar-pdf-set", async (path) => {
 
             if ( ! path.endsWith(".pdf")) {
                 return;
@@ -26,11 +30,25 @@ describe('PDF Metadata', function() {
             const pdfMeta = await PDFMetadata.getMetadata(path);
 
             hitMap.registerHits(...Object.keys(pdfMeta.props));
+            ++nrFiles;
 
         });
 
-        console.log("printing result");
-        console.log(hitMap.toRanked());
+        const percRanked = hitMap.toPercRanked(nrFiles);
+
+        for (const current of percRanked) {
+
+            const strIdx = Strings.lpad(current.idx, ' ', 4);
+            const strKey = Strings.lpad(current.key, ' ', 45);
+            const strPerc = Strings.lpad(current.perc, ' ', 4);
+            const strHits = Strings.lpad(current.hits, ' ', 10);
+
+            console.log(`${strIdx} ${strKey} ${strPerc} ${strHits}`);
+
+        }
+
+        // console.log("printing result");
+        // console.log(hitMap.toRanked());
 
     });